requirements.txt
Django>=3.2,<4.0
gunicorn>=20.1,<21.0
psycopg2>=2.9,<3.0
catboost==1.2.7
joblib==1.4.2
djangorestframework>=3.12,<4.0
pandas>=1.1.0
numpy>=1.19.0
psycopg2>=2.9.10
dotenv>=0.9.9
djangorestframework==3.14.0
django-cors-headers==4.4.0
